Move UserID value object from vo package to userdomain package
