+++ user-rest-service/domain/userdomain/user_id.go
@@ -1,4 +1,4 @@
-package vo
+package userdomain
 
 import (
 	"strings"
